2 interfaces ware added, 2 were changed

+++ field_logic/intarfaces/iDot.java
@@ -1,7 +1,7 @@
 package field_logic.intarfaces;
 
 public interface iDot {
-    double getX();
+    int getX();
 
-    double getY();
+    int getY();
 }
